Update overlay_graph_theta2theta.py
Updated the graph so that I could have subplots of the two samples EH01 and EH02 to make room for the annealed samples on the graph, so that the affect of annealing can be observed.  Also corrected an error I missed in the previous section where I've actually been using the 2theta data for my theta2theta scans, because when I start writing the code I always copy and pasted from previously successfully written scans and forgot to change one section where it's been referring back to the 2theta data.